Fixing jira issue 1285

+++ axiom/src/attachments/mime_parser.c
@@ -1748,7 +1748,7 @@ static axis2_char_t *
                 type++;
             }
             end = type;
-            while (end && *end && !isspace(*end))
+            while (end && *end && !isspace((int)*end))
             {
                 end++;
             }
